[MMG_CRPG] [Client + WebAPI] 유저 보유 캐릭터 장비 #4  아이템 장착 창에서 보유 아이템 보여주기

+++ WebServer/Application/StatTypes/Dtos.cs
@@ -6,7 +6,7 @@ using System.Threading.Tasks;
 
 namespace Application.StatTypes
 {
-    public sealed record StatTypeDto(short Id, string Code, string Name, bool IsPercent);
+    public sealed record StatTypeDto(int Id, string Code, string Name, bool IsPercent);
 
     public sealed record CreateStatTypeRequest(string Code, string Name, bool IsPercent);
     public sealed record UpdateStatTypeRequest(string? Code, string? Name, bool? IsPercent);
